hardware/requirements.txt
aiomqtt==2.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncio-mqtt==0.16.2
asyncpg==0.30.0
click==8.1.8
colorzero==2.0
fastapi==0.115.12
gpiod==2.3.0
gpiodevice==0.0.5
gpiozero==2.0.1
h11==0.16.0
idna==3.10
numpy==2.2.5
paho-mqtt==2.1.0
pigpio==1.78
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
spidev==3.6
st7789==1.0.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
